# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Innowise E-Commerce API",
        "description": "API documentation for Innowise E-Commerce Solutions",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Main API Server"
        }
    ],
    "paths": {
...
        "/api/recommendations/home": {
            "get": {
                "tags": [
                    "Recommendations"
                ],
                "summary": "Get personalized home page recommendations",
                "description": "Returns a personalized list of recommended products based on recently viewed items or random products for new visitors.",
                "operationId": "21e8f67749395f484a4302695922cc6a",
                "parameters": [
                    {
                        "name": "X-Session-Id",
                        "in": "header",
                        "description": "Anonymous visitor session ID (used for tracking without login)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of recommended products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProductResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/{id}/recommendations": {
            "get": {
                "tags": [
                    "Recommendations"
                ],
                "summary": "Get recommendations for a specific product",
                "description": "Returns \"also bought\", \"similar\" and \"recently viewed\" product lists for a given product.",
                "operationId": "cbf888d5b81f2a1034eddfa231c2e79e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "X-Session-Id",
                        "in": "header",
                        "description": "Anonymous visitor session ID",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product recommendations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "also_bought": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "similar": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "recently_viewed": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
        "/api/products/{id}/view": {
            "post": {
                "tags": [
                    "Recommendations"
                ],
                "summary": "Record a product view for recommendation tracking",
                "operationId": "090b843bd1e6455f334b63daddf88ef9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "View recorded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "recorded": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProductResource": {
                "description": "Product resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "category_name": {
                        "type": "string",
                        "example": "Electronics"
                    },
                    "name": {
                        "type": "string",
                        "example": "iPhone 15 Pro"
                    },
                    "description": {
                        "type": "string",
                        "example": "The latest Apple flagship phone."
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 999.99
                    },
                    "old_price": {
                        "type": "number",
                        "format": "float",
                        "example": 1099.99,
                        "nullable": true
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 50
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "url": {
                                    "type": "string",
                                    "format": "uri",
                                    "example": "https://cdn.example.com/products/iphone15.jpg"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}